requirements.txt
pandas
numpy
tqdm
pydantic
pyyaml
flask
flask_cors
requests